# doc/openapi.yaml
openapi: 3.1.0
info:
  title: Vectis HTTP API
  version: v1
  description: >
    Vectis protects data throughout its lifecycle. The API creates key material,
    validates keys, publishes public keys, signs message hashes, exchanges
    protected messages between Vectis instances, and encrypts/decrypts internal
    messages.
servers:
  - url: http://127.0.0.1:3000
    description: Local default server over HTTP
  - url: https://127.0.0.1:3000
    description: Local HTTPS server when VECTIS_MODE=prod
...
security: []
paths:
...
  /self-test/init:
    get:
      tags: [Self Test]
      summary: Validate init key material
      security:
        - ApiKeyAuth: []
      responses:
        "200":
          description: Init key material validation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KeyValidationOutput"
        default:
          $ref: "#/components/responses/Error"
  /self-test/keys/{kid}:
    get:
      tags: [Self Test]
      summary: Validate an operational key
      security:
        - ApiKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/Kid"
      responses:
        "200":
          description: Operational key validation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KeyValidationOutput"
        default:
          $ref: "#/components/responses/Error"
...
  /sign/{kid}:
    post:
      tags: [Signatures]
      summary: Sign a message hash with EdDSA and ML-DSA
      security:
        - ApiKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/Kid"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignInput"
      responses:
        "200":
          description: Compact hybrid signature token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompactSignatureToken"
        "413":
          $ref: "#/components/responses/RequestBodyTooLarge"
        default:
          $ref: "#/components/responses/Error"
  /sign/verification:
    post:
      tags: [Signatures]
      summary: Verify a compact hybrid signature token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompactSignatureToken"
      responses:
        "200":
          description: Signature verification result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerificationOutput"
        "413":
          $ref: "#/components/responses/RequestBodyTooLarge"
        default:
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  parameters:
    Kid:
      name: kid
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/Kid"
  responses:
    Error:
      description: Public error response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    RequestBodyTooLarge:
      description: Request body exceeds the internal 2 MiB limit
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: request body exceeds maximum allowed size
  schemas:
    Kid:
      type: string
      pattern: "^[0-9a-fA-F]{64}$"
      example: f55f086e75b58ac4dfaffd3e75c90d25719281df90e87880145fb9f2e32f2eed
    HexString:
      type: string
      pattern: "^[0-9a-fA-F]+$"
    TextField:
      type: string
      minLength: 1
      pattern: '^[^\x00-\x1F\x7F-\x9F]+$'
...
    Timestamp:
      type: string
      pattern: "^[0-9]+$"
      example: "1782058090"
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          $ref: "#/components/schemas/TextField"
          example: invalid request
...
    KeyValidationOutput:
      type: object
      required: [timestamp, aad, hash, symmetric, eddsa, xecdh, ml-dsa, ml-kem]
      properties:
        timestamp:
          $ref: "#/components/schemas/Timestamp"
        aad:
          $ref: "#/components/schemas/TextField"
        hash:
          type: object
          required: [variant, value_hex]
          properties:
            variant:
              $ref: "#/components/schemas/HashAlgorithm"
            value_hex:
              $ref: "#/components/schemas/HexString"
        symmetric:
          $ref: "#/components/schemas/VariantValidation"
        eddsa:
          $ref: "#/components/schemas/VariantValidation"
        xecdh:
          $ref: "#/components/schemas/VariantValidation"
        ml-dsa:
          $ref: "#/components/schemas/VariantValidation"
        ml-kem:
          $ref: "#/components/schemas/VariantValidation"
    VariantValidation:
      type: object
      required: [variant, valid]
      properties:
        variant:
          $ref: "#/components/schemas/TextField"
        valid:
          type: boolean
    SignInput:
      type: object
      additionalProperties: false
      required: [message_hash]
      properties:
        message_hash:
          $ref: "#/components/schemas/MessageHash"
    MessageHash:
      oneOf:
        - $ref: "#/components/schemas/Hash160"
        - $ref: "#/components/schemas/Hash224Blake2b"
        - $ref: "#/components/schemas/Hash224Sha2"
        - $ref: "#/components/schemas/Hash224Sha3"
        - $ref: "#/components/schemas/Hash256"
        - $ref: "#/components/schemas/Hash384"
        - $ref: "#/components/schemas/Hash512"
    Hash160:
      type: object
      additionalProperties: false
      required: [alg, hex]
      properties:
        alg:
          const: BLAKE2b(160)
        hex:
          type: string
          pattern: "^[0-9a-fA-F]{40}$"
    Hash224Blake2b:
      type: object
      additionalProperties: false
      required: [alg, hex]
      properties:
        alg:
          const: BLAKE2b(224)
        hex:
          type: string
          pattern: "^[0-9a-fA-F]{56}$"
    Hash224Sha2:
      type: object
      additionalProperties: false
      required: [alg, hex]
      properties:
        alg:
          const: SHA-224
        hex:
          type: string
          pattern: "^[0-9a-fA-F]{56}$"
    Hash224Sha3:
      type: object
      additionalProperties: false
      required: [alg, hex]
      properties:
        alg:
          const: SHA-3(224)
        hex:
          type: string
          pattern: "^[0-9a-fA-F]{56}$"
    Hash256:
      type: object
      additionalProperties: false
      required: [alg, hex]
      properties:
        alg:
          enum: [BLAKE2b(256), SHA-256, SHA-512-256, SHA-3(256)]
        hex:
          type: string
          pattern: "^[0-9a-fA-F]{64}$"
    Hash384:
      type: object
      additionalProperties: false
      required: [alg, hex]
      properties:
        alg:
          enum: [BLAKE2b(384), SHA-384, SHA-3(384)]
        hex:
          type: string
          pattern: "^[0-9a-fA-F]{96}$"
    Hash512:
      type: object
      additionalProperties: false
      required: [alg, hex]
      properties:
        alg:
          enum: [BLAKE2b(512), SHA-512, SHA-3(512), Whirlpool]
        hex:
          type: string
          pattern: "^[0-9a-fA-F]{128}$"
    CompactSignatureToken:
      type: object
      additionalProperties: false
      required: [kid, signature]
      properties:
        kid:
          $ref: "#/components/schemas/Kid"
        signature:
          type: string
          maxLength: 65536
          description: "Four unpadded Base64URL segments: header.payload.eddsa.ml-dsa."
...
    VerificationOutput:
      type: object
      required: [status, valid]
      properties:
        status:
          type: object
          required: [eddsa, ml-dsa]
          properties:
            eddsa:
              enum: [ok, fail, not_checked]
            ml-dsa:
              enum: [ok, fail]
        valid:
          enum: [ok, fail]
...
    HashAlgorithm:
      enum:
        - BLAKE2b(160)
        - BLAKE2b(224)
        - BLAKE2b(256)
        - BLAKE2b(384)
        - BLAKE2b(512)
        - SHA-224
        - SHA-256
        - SHA-384
        - SHA-512
        - SHA-512-256
        - SHA-3(224)
        - SHA-3(256)
        - SHA-3(384)
        - SHA-3(512)
        - Whirlpool
...