# src/io/http/sign.rs
use super::HttpState;
use super::error::{ErrorResponse, error_response};
use super::extract::JsonBody;
use crate::core::{audit, blocking, metrics};
use crate::ops;
use axum::Json;
use axum::extract::{Path, State};
use axum::http::{HeaderMap, StatusCode};
use tracing::{error, info};

pub async fn sign_endpoint(
    State(state): State<HttpState>,
    Path(id): Path<String>,
    headers: HeaderMap,
    JsonBody(request): JsonBody,
) -> Result<Json<ops::sign::CompactSignatureToken>, (StatusCode, Json<ErrorResponse>)> {
    let request_context = state.authorize_request(&headers).await?;
    request_context.require_permission_for(Some(&id), "sign", Some("sign.denied"))?;
    let actor = audit::actor_from_client(request_context.client());

    ops::keys::validate_key_id(&id).map_err(|err| {
        audit::operation_failed(
            "sign.failed",
            Some(&actor),
            Some(&id),
            None,
            Some("sign"),
            &err.to_string(),
        );
        metrics::record_crypto_operation("sign", "failed");
        error_response(err.as_ref())
    })?;
    state.ensure_keys_db_entry(&id).await.map_err(|err| {
        audit::operation_failed(
            "sign.failed",
            Some(&actor),
            Some(&id),
            None,
            Some("sign"),
            &err.to_string(),
        );
        metrics::record_crypto_operation("sign", "failed");
        error_response(err.as_ref())
    })?;
    let request = ops::sign::parse_sign_input(request).map_err(|err| {
        audit::operation_failed(
            "sign.failed",
            Some(&actor),
            Some(&id),
            None,
            Some("sign"),
            &err.to_string(),
        );
        metrics::record_crypto_operation("sign", "failed");
        error_response(err.as_ref())
    })?;
    info!(
        endpoint = "/sign/{kid}",
        kid = %id,
        hash_alg = %request.message_hash.alg,
        hash_hex_len = request.message_hash.hex.len(),
        "sign request accepted"
    );

    let loaded_key = state
        .with_keys_db_state(|keys_db_state| ops::keys::get_loaded_key(keys_db_state, &id))
        .await
        .map_err(|err| {
            audit::operation_failed(
                "sign.failed",
                Some(&actor),
                Some(&id),
                None,
                Some("sign"),
                &err.to_string(),
            );
            metrics::record_crypto_operation("sign", "failed");
            error_response(err.as_ref())
        })?;
    let result = blocking::spawn_blocking_crypto(move || {
        ops::sign::sign_timestamp_with_loaded_key(&loaded_key, request)
    })
    .await;

    match result {
        Ok(response) => {
            info!(
                endpoint = "/sign/{kid}",
                kid = %response.kid,
                signature_len = response.signature.len(),
                "sign response ready"
            );
            audit::operation_success(
                "sign.success",
                Some(&actor),
                Some(&response.kid),
                None,
                Some("sign"),
            );
            metrics::record_crypto_operation("sign", "success");

            Ok(Json(response))
        }
        Err(err) => {
            audit::operation_failed(
                "sign.failed",
                Some(&actor),
                Some(&id),
                None,
                Some("sign"),
                &err.to_string(),
            );
            metrics::record_crypto_operation("sign", "failed");
            error!(error = %err, id = %id, "sign endpoint failed");
            Err(error_response(err.as_ref()))
        }
    }
}

pub async fn sign_verification_endpoint(
    State(state): State<HttpState>,
    JsonBody(request): JsonBody,
) -> Result<Json<ops::sign::VerificationOutput>, (StatusCode, Json<ErrorResponse>)> {
    let config = state.config_snapshot().await;
    let request = ops::sign::parse_compact_signature_token(request).map_err(|err| {
        audit::operation_failed("verify.failed", None, None, None, None, &err.to_string());
        metrics::record_crypto_operation("verify", "failed");
        error_response(err.as_ref())
    })?;
    let kid = request.kid.clone();
    info!(
        endpoint = "/sign/verification",
        kid = %kid,
        signature_len = request.signature.len(),
        "sign verification request accepted"
    );

    let result = match state.ensure_keys_db_entry(&kid).await {
        Ok(()) => {
            let loaded_key = state
                .with_keys_db_state(|keys_db_state| ops::keys::get_loaded_key(keys_db_state, &kid))
                .await;
            match loaded_key {
                Ok(loaded_key) => {
                    blocking::spawn_blocking_crypto(move || {
                        ops::sign::verify_compact_timestamp_with_loaded_key(&loaded_key, &request)
                    })
                    .await
                }
                Err(err) => Err(err),
            }
        }
        Err(local_err) => match config.remote_routes.public_keys_for(&kid).cloned() {
            Some(peer) => {
                blocking::spawn_blocking_crypto(move || {
                    ops::sign::verify_compact_timestamp_with_peer_keys(&request, &peer)
                })
                .await
            }
            None => Err(local_err),
        },
    };

    match result {
        Ok(response) => {
            info!(
                endpoint = "/sign/verification",
                kid = %kid,
                valid = %response.valid,
                eddsa = %response.status.eddsa,
                ml_dsa = %response.status.ml_dsa,
                "sign verification response ready"
            );
            audit::operation_success("verify.success", None, Some(&kid), None, None);
            metrics::record_crypto_operation("verify", "success");

            Ok(Json(response))
        }
        Err(err) => {
            audit::operation_failed(
                "verify.failed",
                None,
                Some(&kid),
                None,
                None,
                &err.to_string(),
            );
            metrics::record_crypto_operation("verify", "failed");
            error!(error = %err, kid = %kid, "sign verification endpoint failed");
            Err(error_response(err.as_ref()))
        }
    }
}